requirements.txt
aiogram==3.1.1
aiohttp~=3.8.6
python-dotenv==1.0.1
pydantic>=2.1.1,<2.4
